ui/requirements.txt
huggingface_hub==0.25.2
llama-cpp-python
transformers[torch]
torchaudio